requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
ollama
faiss-cpu
neo4j
sentence-transformers
